Add a database view to query all agencies of a req.
Requirements are connected to individual agencies as well as agency groups.
We'd like to find *all* agencies associated with a requirement, including
through the agencies. To do this, we can use a database view that unions the
direct connection and indirect connection.

As we're using a view, though, we need to prevent the data from being editable
in the admin.