Remove automatic redirect when searching.
For folks without JS, we had been first hitting the API for an exact text
match (we'd redirect if one existed). That's neat, but makes the logic a bit
painful. Further, it won't really work if the user is searching over multiple
fields (like the abbr and name fields of an agency.)

Remove that capability.